+++ src/main/java/com/matias/springjwt/repository/IUserRepository.java
@@ -7,9 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IUserRepository extends JpaRepository<UserEntity, Long> {
+public interface IUserRepository extends JpaRepository<UserEntity, Integer> {
     
     Optional<UserEntity> findByUsername(String username);
 
     Boolean existsByUsernameOrEmail(String username, String email);
+    
 }
